TARGET clinical merge raw tables

+++ BQ_Table_Building/Clinical_From_File/build_gdc_file_clinical_tables_TARGET.py
@@ -21,6 +21,7 @@ import shutil
 import logging
 import sys
 import time
+from decimal import Decimal
 
 import pandas as pd
 
@@ -470,7 +471,7 @@ def main(args):
                         if value == '.' or value is None:
                             value = None
                         else:
-                            value = float(value)
+                            value = Decimal(value)
 
                     if value is None:
                         continue
